MSR config/results through kdb+
The MSR config driver can now be controlled through kdb+. The list of performance metrics is loaded by kdb+ from a csv file (pmcdata.csv).
Functions are provided to configure which performance events are counted.
The shared library configures the MSRs and executes the test code, and collates the results in kdb+ int vectors. These can easily be turned into a table in the kdb+ environment.
Since kdb+ is scriptable, successive runs over the same code (with different perf counters) can easily be executed and analysed.