Initial code
This is a toy implementaiton I'm writing to get me used to working with
structs, the heap, arrays and pointers. It could be useful if it ever
becomes full-featured with things like remove or iterate...

ATM it's only got put and get (and no rehash or growth facility). It
probably should be adaptable to any type.